requirements.txt
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
daphne
channels
channels-redis
pillow
whitenoise
dj-database-url
psycopg2-binary
python-dotenv
gunicorn
async-timeout
celery
redis
django-modeltranslation
